Refactor common functions
